Dissect login.ts into multiple files
